Add delay before add and implement rotation on edges
When element is about the be added on board i.e. it can
not go lower, player should be left with slight delay so additional
move can be made to find better fit or even make element
enabled to go lower.

Also, when element is on far left or far right of the board and
player wants to rotate it, it should be moved toward center
of the board and rotation should be performed, if it is
possible.